requirements.txt
absl-py==1.2.0
accelerate==0.23.0
aiohttp==3.9.1
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
clip==0.2.0
cmake==3.27.2
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.1.1
cryptography==41.0.7
cycler==0.11.0
decorator==5.1.1
diffusers==0.24.0
docker-pycreds==0.4.0
einops==0.7.0
executing==1.2.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.44.3
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
gitdb==4.0.10
gitpython==3.1.38
google-auth==2.25.2
google-auth-oauthlib==1.0.0
grpcio==1.60.0
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
imhist==0.0.4
importlib-metadata==6.8.0
importlib-resources==5.12.0
imwatermark==0.0.2
invisible-watermark==0.1.5
ipython==8.10.0
jedi==0.18.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.4
kornia==0.6.12
lazy-loader==0.2
lightning-utilities==0.10.0
markdown==3.5.1
markupsafe==2.1.1
matplotlib==3.7.2
matplotlib-inline==0.1.3
ml-collections==0.1.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.12.0
onnxruntime==1.16.2
open-clip-torch==2.23.0
opencv-python==4.8.0.76
packaging==23.2
pandas==2.0.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.0.1
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.38
protobuf
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeprecate==0.3.2
pygments==2.14.0
pyopenssl==23.2.0
pyparsing==3.0.9
pyproject-toml==0.0.11
pysocks==1.7.1
python-dateutil==2.8.2
pytorch-fid==0.2.1
pytorch-lightning==1.6.5
pywavelets==1.4.1
pyyaml==6.0.1
readline==6.2.4.1
referencing==0.32.0
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.11.2
sentencepiece==0.1.99
sentry-sdk==1.24.0
setproctitle==1.3.2
setuptools==68.0.0
six==1.16.0
smmap==5.0.1
stack-data==0.6.2
sympy==1.12
taming-transformers==0.0.1
taming-transformers-rom1504==0.0.6
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tifffile==2023.7.10
timm==0.4.5
tokenizers==0.13.3
toml==0.10.2
torchvision==0.19.0+cu121
tqdm==4.66.1
traitlets==5.9.0
transformers==4.33.3
typing_extensions>=4.8.0,<5.0.0
tzdata==2023.4
urllib3==1.26.18
wandb==0.15.3
wcwidth==0.2.12
werkzeug==3.0.1
wheel==0.41.2
yarl==1.9.4
zipp==3.15.0
